Extract DC inferring logic to separate hierarchy
