+++ core/src/main/java/com/datastax/oss/driver/internal/core/loadbalancing/DcInferringLoadBalancingPolicy.java
@@ -18,16 +18,10 @@ package com.datastax.oss.driver.internal.core.loadbalancing;
 import com.datastax.oss.driver.api.core.config.DefaultDriverOption;
 import com.datastax.oss.driver.api.core.context.DriverContext;
 import com.datastax.oss.driver.api.core.loadbalancing.LoadBalancingPolicy;
-import com.datastax.oss.driver.api.core.metadata.Node;
-import com.datastax.oss.driver.internal.core.metadata.DefaultNode;
+import com.datastax.oss.driver.internal.core.loadbalancing.helper.InferringLocalDcHelper;
+import com.datastax.oss.driver.internal.core.loadbalancing.helper.LocalDcHelper;
 import edu.umd.cs.findbugs.annotations.NonNull;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
 import net.jcip.annotations.ThreadSafe;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * An implementation of {@link LoadBalancingPolicy} that infers the local datacenter from the
@@ -71,72 +65,13 @@ import org.slf4j.LoggerFactory;
 @ThreadSafe
 public class DcInferringLoadBalancingPolicy extends BasicLoadBalancingPolicy {
 
-  private static final Logger LOG = LoggerFactory.getLogger(DcInferringLoadBalancingPolicy.class);
-
   public DcInferringLoadBalancingPolicy(
       @NonNull DriverContext context, @NonNull String profileName) {
     super(context, profileName);
   }
 
-  /**
-   * Discovers the local datacenter to use with this policy.
-   *
-   * <p>This method should be called upon {@linkplain #init(Map, DistanceReporter) initialization}.
-   *
-   * <p>This implementation fetches the user-supplied datacenter, if any, from the programmatic
-   * configuration API, or else, from the driver configuration. If no local datacenter is explicitly
-   * defined, this implementation infers the local datacenter from the contact points: if all
-   * contact points share the same datacenter, that datacenter is returned. If the contact points
-   * are from different datacenters, or if no contact points reported any datacenter, an {@link
-   * IllegalStateException} is thrown.
-   *
-   * @return The local datacenter; always present.
-   * @throws IllegalStateException if the local datacenter could not be inferred.
-   */
-  @NonNull
   @Override
-  protected Optional<String> discoverLocalDatacenter() {
-    Optional<String> optionalLocalDc = super.discoverLocalDatacenter();
-    if (optionalLocalDc.isPresent()) {
-      return optionalLocalDc;
-    }
-    return inferDataCenter(this);
-  }
-
-  /**
-   * DC inference logic, exposed as a static utility method for convenience, to allow for easier
-   * subclassing.
-   *
-   * @return The local datacenter; always present.
-   * @throws IllegalStateException if the local datacenter could not be inferred.
-   */
-  @NonNull
-  public static Optional<String> inferDataCenter(BasicLoadBalancingPolicy policy) {
-    Set<String> datacenters = new HashSet<>();
-    Set<DefaultNode> contactPoints = policy.context.getMetadataManager().getContactPoints();
-    for (Node node : contactPoints) {
-      String datacenter = node.getDatacenter();
-      if (datacenter != null) {
-        datacenters.add(datacenter);
-      }
-    }
-    if (datacenters.size() == 1) {
-      String localDc = datacenters.iterator().next();
-      LOG.info("[{}] Inferred local DC from contact points: {}", policy.logPrefix, localDc);
-      return Optional.of(localDc);
-    }
-    if (datacenters.isEmpty()) {
-      throw new IllegalStateException(
-          "The local DC could not be inferred from contact points, please set it explicitly (see "
-              + DefaultDriverOption.LOAD_BALANCING_LOCAL_DATACENTER.getPath()
-              + " in the config, or set it programmatically with SessionBuilder.withLocalDatacenter)");
-    }
-    throw new IllegalStateException(
-        String.format(
-            "No local DC was provided, but the contact points are from different DCs: %s; "
-                + "please set the local DC explicitly (see "
-                + DefaultDriverOption.LOAD_BALANCING_LOCAL_DATACENTER.getPath()
-                + " in the config, or set it programmatically with SessionBuilder.withLocalDatacenter)",
-            policy.formatNodes(contactPoints)));
+  protected LocalDcHelper createLocalDcHelper() {
+    return new InferringLocalDcHelper(context, profile, logPrefix);
   }
 }
