+++ core/src/main/java/com/datastax/oss/driver/internal/core/loadbalancing/helper/LocalDcHelper.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright DataStax, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.datastax.oss.driver.internal.core.loadbalancing.helper;
+
+import java.util.Optional;
+
+public interface LocalDcHelper {
+
+  /**
+   * Returns the local datacenter, if it can be discovered, or returns {@link Optional#empty empty}
+   * otherwise.
+   *
+   * <p>Implementors may choose to throw {@link IllegalStateException} instead of returning {@link
+   * Optional#empty empty}, if they require a local datacenter to be defined in order to operate
+   * properly.
+   *
+   * @return The local datacenter, or {@link Optional#empty empty} if none found.
+   * @throws IllegalStateException if the local datacenter could not be discovered, and this policy
+   *     cannot operate without it.
+   */
+  Optional<String> discoverLocalDc();
+}
